requirements.txt
Flask
rembg
Pillow
Jinja2
gunicorn
